requirements.txt
matplotlib==3.10.0
mung==1.2.1
numpy==2.2.2
opencv_python==4.10.0.84
packaging==24.2
Pillow==11.1.0
prettytable==3.14.0
PyYAML==6.0.2
requests==2.32.3
seaborn==0.13.2
smashcima==0.2.0
tqdm==4.66.1
ultralytics==8.3.28
